Prijs van decimal -> int

+++ src/App/Domain/Bod.cs
@@ -9,11 +9,11 @@ namespace Domain {
     public class Bod {
         public int Id { get; private set; }
         public Klant Klant { get; private set; }
-        public decimal BodPrijs { get; set; }
+        public int BodPrijs { get; set; }
         public DateTime Datum { get; set; }
         public int KlantId { get; private set; }
 
-        public Bod(Klant klant, decimal bodPrijs, DateTime datum) {
+        public Bod(Klant klant, int bodPrijs, DateTime datum) {
             Klant = Guard.Against.Null(klant, nameof(klant));
             KlantId = klant.GebruikerId;
             BodPrijs = Guard.Against.Null(bodPrijs, nameof(bodPrijs));
