+++ src/App/Project3H04/Shared/DTO/Bod_DTO.cs
@@ -8,7 +8,7 @@ using Project3H04.Shared.Veilingen;
 namespace Project3H04.Shared.DTO {
     public class Bod_DTO {
         public Klant_DTO Klant { get; set; }
-        public decimal BodPrijs { get; set; }
+        public int BodPrijs { get; set; }
         public DateTime Datum { get; set; }
     }
 }
